// test/integration/test/ports/_fixtures/naming-conflict-model-vs-model/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
type DefaultLiteralValue<CodecId extends string, _Encoded> = CodecId extends keyof CodecTypes
  ? CodecTypes[CodecId]['output']
  : _Encoded;

export type FieldOutputTypes = {
  readonly public: {
    readonly Model: { readonly id: Char<24>; readonly otherId: CodecTypes['pg/text@1']['output'] };
    readonly ModelUpdate: {
      readonly id: Char<24>;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Model: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly otherId: CodecTypes['pg/text@1']['input'];
    };
    readonly ModelUpdate: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly model: { readonly id: Char<24>; readonly otherId: CodecTypes['pg/text@1']['output'] };
    readonly modelUpdate: {
      readonly id: Char<24>;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly model: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly otherId: CodecTypes['pg/text@1']['input'];
    };
    readonly modelUpdate: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;